Cells: [rax_only] Server migration (nova migrate) fix.
Migration was failing on Glance image metadata lookups.  The problem was
that Racker admin auth tokens are only capable of working with specific
Glance API/registry servers allocated for Racker admins.  When a resize
or migrate request was forwarded to a child cell configured to utilize
a different pool of Glance servers, the auth token was rejected.

This fix allows the client to optionally tag the context with a list
of Glance API servers to use.

Change-Id: I4328118ab04ee78a6ca361d12c5eb26027e7f9a4

Ply-Patch: Cells-rax_only-Server-migration-nova-migrate-fix.patch