Allow revert resize from ERROR
Fixes D-08031

If an instance fails to resize, the instance is moved to error with
neither the old nor new instances running. The old instance often
can be restarted, but the OpenStack Compute API complicates that.
As a workaround, allow revert resize from ERROR if the task_state is
set correctly to indicate it failed during a resize.