Remove bandwidth data from immediate exists events.
When we were polling XenServer RRD's for bandwidth, it would zero the
bandwidth used on resize/rebuilds. Since nova itself is now doing the
polling of the VIFs, this no longer happens.

Usage was previously adding together bandwidth from the immediate .exists
events emitted on resize/rebuilds with the value from the end-of-day events
to compensate. That is no longer needed, so we need to remove the bandwidth
data from the immediate events to prevent overbilling.

Change-Id: If5ec1fe715cfe1145d9382108a5824f558446ed6

Fix collision between conductor and local bw patch.

Change 1df9f6a2d717e9249aab9c61448d2b11442d4803 added a patch
to remove bw data from the 'immediate' .exists events in our
development branch (for billing). When the conductor code landed
in trunk, it changed the way notify_usage_exists was being called
(now it's called through the conductor api) and this was breaking
the instance_usage_audit periodic task that eimits end-of-day
usage events, because there was an extra argument the conductor
code wasn't expecting.

This fixes the conductor api to work with our local patch.

Change-Id: I4bda6557b5c1bf3ec2b45c0dff6519a2cde6a93c

Ply-Patch: Remove-bandwidth-data-from-immediate-exists-events.patch