Implements Scheduled Images Extension
The extension creates a new resource under servers/id to get, create
and delete it's scheduled image.

On the servers list and detail calls, it adds the system metadata
related to scheduled image service on the server dict in the response
object. It also enables filtering by the system metadata added by the
scheduled images service and if it exists, adds it on the servers
dictonaries.

bp https://blueprints.launchpad.net/nova/+spec/scheduled-images

Squashed:

Add username to schedule image metadata for auth impersonation

Squashed:

SI ext makes compute api calls.

scheduled images ext makes compute api calls instead of db for
operations on the system metadata.

Change-Id: I72da0e9c32f7d5586754bfad56ecd19ed3200b08

Squashed:

Add authorize check for instances in scheduled_images ext

Squashed:

Avoid un-necessary QonoS calls on server DELETE

Currently, QonoS API is being called for all the servers to check if the
schedule exists for that server. In this fix, QonoS will be called only
for those servers for which the retention metadata has been set on them.

Squashed:

Fixes line too long pep8 errors in sched imgs ext

A couple of LOG lines are way too long after indentation which is
fixed in this commit.

Squashed:

Change os-si-image-schedule extension to rax-si-image-schedule

Squashed:

Allow SI to be turned off if schedule missing

If things get out of sync and a schedule is deleted while the
instance still has retention set, a user cannot disable or
re-enable the schedule. Do not raise an error when deleting a
scheduled if it's already gone.

Addresses: RM4650

Ply-Patch: Implements-Scheduled-Images-Extension.patch