API to reset a db state of a migration
Adding a new extension to reset_migration_state to fix failed migrations
When a migration hangs or fails, the migration is manually recovered.
This API is called to reset the migration and instance states to mark the
migration as complete or to revert the migration.

Change-Id: Ia20ae3c4b34788d4ed3c44637a1fdde5ed24ff39
Implements: blueprint revert-resize-force-option

Ply-Patch: API-to-reset-a-db-state-of-a-migration.patch