Don't deallocate/reallocate networks on reschedules
If the virt driver doesn't have dhcp_options_for_instance (only
baremetal seems to) then network allocation isn't dependent upon the
host the instance is on.  In that case we don't need to deallocate and
reallocate a network for each reschedule.

The current method in the code which checks for the length of the list
returned from _get_instance_nw_info() doesn't work because it can't
distinguish between an unallocated network and one that doesn't return
network information.

Part of bp remove-cast-to-schedule-run-instance

Change-Id: Ic427715f46cb5179a37c593c4217f239c65544c3