Use extra_resources field in filter scheduler
The extra_resources field in the compute_nodes table allows a compute
node to pass an extensible set of information about resource
availability at a compute node. This is used in the filter scheduler
as an additional source of information on which to base scheduling
decisions.

In addition, the consumer plug-in is added to the filter scheduler.
When the scheduler has selected a compute node it updates its local
copy of HostState for that compute node to reflect the decision.
The consumer plug-in provides the operation required on the extended
resource information to perform this step.

This is part of: blueprint extensible-resource-tracking

Change-Id: I9a206808f4b60118a394f9148a8fcf20cd46720a