Add image build filter to Nova
Some images can't be used to boot an instance based on a combination of
user roles and image metadata.  This is enforced by Glance in a way hat
causes some discrepencies betwen image list and image show.  The proper
place for this enforcement is Nova, so this ports the logic from Glance
into Nova.

https://redmine.ohthree.com/issues/4344

Change-Id: Ife9875ef519790bdbaeb2691de0602f0356b3817

Ply-Patch: Add-image-build-filter-to-Nova.patch