Fix vif ordering in Neutron
RM5616

Don't sort a single vif

No need to attempt to sort since its a dict and its only going to be a
single element list anyway.

RM5660

Duplicate CONF options so they are registered

network_order was only being registered in the quantum2 manager.
Duplicate that option in the assignable api and migrate the ordering to
it to reduce surface area.

RM5666

Change-Id: I6ec6b1151e3f8450cdcf9f4771369a2b2e002768

Ply-Patch: Fix-vif-ordering-in-Neutron.patch