Use cloudBlockStorageRackerAdmin volume endpoint if set
When a request is made through the Nova racker admin API the service
catalog returned will include a CBS racker admin api endpoint. In that
case we want cinder requests routed to there because they share an auth
system so tokens will validate. We should check for the presence of hat
endpoint and use it when it is set.

https://redmine.ohthree.com/issues/5848

Change-Id: I8e7a07be770a74c2025f3b877153714c9c938ff8

Ply-Patch: Use-cloudBlockStorageRackerAdmin-volume-endpoint-if-.patch