Reap idle SQL connections after 60 seconds
Change the amount of time required before idle SQL connections are
removed from the DB connection pool and closed.

This value was previously set to an hour, which had the potential to
overload a DB server with an unnecessary amount of idle connections
depending on the number of nova computes and control plane nodes.

UpgradeImpact Reduce default time to reap idle SQL connections from 1 hour to
60 seconds.

Change-Id: Ibe9cc18999e3320379a6f9e0b86b024ba3ea4750